Added in a basic thread posting front-end layout
